+++ procurement_server/src/main/java/com/procurement/procurement_server/dao/OrderRepo.java
@@ -2,10 +2,9 @@ package com.procurement.procurement_server.dao;
 
 import java.util.List;
 
-import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import com.procurement.procurement_server.model.Order;
+import com.procurement.procurement_server.model.order_level.Order;
 
 
 public interface OrderRepo extends MongoRepository<Order, String>{
