+++ procurement_server/src/main/java/com/procurement/procurement_server/service/ServiceHandler.java
@@ -1,25 +1,22 @@
 package com.procurement.procurement_server.service;
 
-import com.procurement.procurement_server.model.Order;
-import com.procurement.procurement_server.model.Requistion;
-import com.procurement.procurement_server.model.user_level.User;
-import com.procurement.procurement_server.service.user_service.UserService;
 
-import org.bson.types.ObjectId;
+import com.procurement.procurement_server.model.order_level.Order;
+import com.procurement.procurement_server.model.order_level.Requistion;
+import com.procurement.procurement_server.model.user_level.User;
+import com.procurement.procurement_server.service.order_service.*;
+import com.procurement.procurement_server.service.order_service.builder.ApprovedOrder;
+import com.procurement.procurement_server.service.order_service.builder.OrderBroker;
+import com.procurement.procurement_server.service.order_service.builder.OrderBuilder;
+import com.procurement.procurement_server.service.order_service.builder.PendingOrder;
+import com.procurement.procurement_server.service.user_service.UserService;
+import com.procurement.procurement_server.util.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
-import com.procurement.procurement_server.util.ApprovedOrder;
-import com.procurement.procurement_server.util.CommonConstants;
-import com.procurement.procurement_server.util.Generator;
-import com.procurement.procurement_server.util.OrderBroker;
-import com.procurement.procurement_server.util.OrderBuilder;
-import com.procurement.procurement_server.util.PendingOrder;
-
 import java.util.ArrayList;
-import java.util.logging.Handler;
 
 @Component
 public class ServiceHandler {
@@ -31,12 +28,11 @@ public class ServiceHandler {
 
     @Autowired
     DataServer dataServer;
-    
-    @Autowired
-    IOrderService orderService;
 
-    
-    public ResponseEntity<Object> handleServiceRequest(String reqId, Object obj, String uid) {
+    @Autowired
+    OrderServiceImpl orderService;
+
+    public ResponseEntity handleServiceRequest(String reqId, Object obj, String uid) {
         if (!isIsInitialized()) {
             startDataServer();
             setIsInitialized(true);
@@ -47,15 +43,17 @@ public class ServiceHandler {
             case CommonConstants.ADD_USER_REQUEST:
                 return addNewUser(obj);
             case CommonConstants.GET_AVAILABLE_SUPPLIER_ITEMS:
-                return null;
+                return getAvailableItemsList();
             case CommonConstants.GET_ALL_USERS:
                 return getAllUsers();
             case CommonConstants.DELETE_SPECIFIC_USER:
                 return deleteSpecificUser(uid);
             case CommonConstants.ADD_ORDER_REQUEST:
-            	return handleOrder((Order)obj); 
-            case CommonConstants.UPDATE_ORDER_REQUEST : 
+            	return handleOrder((Order)obj);
+            case CommonConstants.UPDATE_ORDER_REQUEST :
             	return handleOrder((Order) obj);
+                case CommonConstants.GET_ALL_ORDERS:
+                    return getAllOrders();
             default:
                 return new ResponseEntity("Failed", HttpStatus.OK);
         }
@@ -83,18 +81,18 @@ public class ServiceHandler {
     }
 
     public ResponseEntity<Object> handleOrder(Order order ) {
-    	
-    
+
+
     		int orderItemQuantity = orderService.calculateQuantity(order.getItems());
     		double orderTotal = orderService.calculateTotal(order.getItems());
-    		
+
     		Requistion requisition = new Requistion();
-    		
-    		
+
+
     		ApprovedOrder approveOrder = new ApprovedOrder(requisition);
     		PendingOrder pendingOrder = new PendingOrder(requisition);
-    		
-    		
+
+
     		if( order.get_idAsObjectId() == null ) {
     			System.out.println("id is null");
     			requisition.set_id(new ObjectId());
@@ -102,11 +100,11 @@ public class ServiceHandler {
     			System.out.println("id is not null");
     			requisition = order.getRequistion();
     		}
-    		
+
     		/*-----------------------------------------------------------*/
-    		
-    		OrderBroker broker = new OrderBroker();		
-    		
+
+    		OrderBroker broker = new OrderBroker();
+
     		if( orderTotal > CommonConstants.ORDER_LIMIT ) {
     			broker.takeOrder(pendingOrder);
     			requisition =  broker.placeOrder();
@@ -114,8 +112,8 @@ public class ServiceHandler {
     			broker.takeOrder(approveOrder);
     			requisition = broker.placeOrder();
     		}
-    		
-    		
+
+
     		Order newOrder = new OrderBuilder(order.get_idAsObjectId())
     				.setItems(order.getItems())
     				.setOrderDate(Generator.getCurrentDate())
@@ -124,13 +122,13 @@ public class ServiceHandler {
     				.setRequisition(null)
     				.setTotalAmount(orderTotal)
     				.build();
-    		
+
     		return new ResponseEntity<>(orderService.addOrder(newOrder, requisition), HttpStatus.OK);
-    	
-	
+
+
     }
     
-    /*
+
     public ResponseEntity getAvailableItemsList() {
         ArrayList<Items> itemsList = new ArrayList<>();
         Items items = new Items();
@@ -160,7 +158,7 @@ public class ServiceHandler {
 
         return new ResponseEntity<Object>(itemsList, HttpStatus.OK);
     }
-    */
+
 
     private ResponseEntity getAllUsers() {
         return userService.getAllUsers();
