--- procurement_server/src/main/java/com/procurement/procurement_server/service/TestService.java
@@ -1,5 +1,0 @@
-package com.procurement.procurement_server.service;
-
-public class TestService {
-
-}
