+++ procurement_server/src/main/java/com/procurement/procurement_server/service/order_service/OrderServiceImpl.java
@@ -1,15 +1,16 @@
-package com.procurement.procurement_server.service;
+package com.procurement.procurement_server.service.order_service;
 
 import java.util.List;
 
+import com.procurement.procurement_server.service.order_service.builder.IOrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.procurement.procurement_server.dao.OrderRepo;
 import com.procurement.procurement_server.dao.RequisitionRepo;
-import com.procurement.procurement_server.model.Order;
-import com.procurement.procurement_server.model.Requistion;
+import com.procurement.procurement_server.model.order_level.Order;
+import com.procurement.procurement_server.model.order_level.Requistion;
 import com.procurement.procurement_server.model.supplier_level.Item;
 
 @Service
