+++ procurement_server/src/main/java/com/procurement/procurement_server/service/order_service/builder/ApprovedOrder.java
@@ -1,8 +1,8 @@
-package com.procurement.procurement_server.util;
+package com.procurement.procurement_server.service.order_service.builder;
 
-import com.procurement.procurement_server.model.Requistion;
+import com.procurement.procurement_server.model.order_level.Requistion;
 
-public class ApprovedOrder implements OrderCommand{
+public class ApprovedOrder implements OrderCommand {
 	
 	private Requistion requisition;
