+++ procurement_server/src/main/java/com/procurement/procurement_server/service/order_service/builder/OrderCommand.java
@@ -1,0 +1,8 @@
+package com.procurement.procurement_server.service.order_service.builder;
+
+import com.procurement.procurement_server.model.order_level.Requistion;
+
+public interface OrderCommand {
+
+	public Requistion execute();
+}
