+++ procurement_server/src/main/java/com/procurement/procurement_server/service/order_service/builder/PendingOrder.java
@@ -1,6 +1,6 @@
-package com.procurement.procurement_server.util;
+package com.procurement.procurement_server.service.order_service.builder;
 
-import com.procurement.procurement_server.model.Requistion;
+import com.procurement.procurement_server.model.order_level.Requistion;
 
 public class PendingOrder implements OrderCommand{
