--- procurement_server/src/main/java/com/procurement/procurement_server/util/OrderCommand.java
@@ -1,8 +1,0 @@
-package com.procurement.procurement_server.util;
-
-import com.procurement.procurement_server.model.Requistion;
-
-public interface OrderCommand {
-
-	public Requistion execute();
-}
